[Fix] 필기 폭을 Double로 저장

+++ semomun/StudyVCs/ChildVCs/FormOne/MultipleWithNoAnswerVC.swift
@@ -244,7 +244,7 @@ extension MultipleWithNoAnswerVC: UICollectionViewDelegateFlowLayout{
 extension MultipleWithNoAnswerVC {
     func canvasViewDrawingDidChange(_ canvasView: PKCanvasView) {
         let data = self.canvasView.drawing.dataRepresentation()
-        self.viewModel?.updatePagePencilData(to: data, width: Int(self.canvasView.contentSize.width))
+        self.viewModel?.updatePagePencilData(to: data, width: Double(self.canvasView.contentSize.width))
     }
 }
