+++ semomun/StudyVCs/ChildVCs/SingleWithNoAnswerVC.swift
@@ -277,7 +277,7 @@ extension SingleWithNoAnswerVC {
 extension SingleWithNoAnswerVC: PKCanvasViewDelegate {
     func canvasViewDrawingDidChange(_ canvasView: PKCanvasView) {
         let data = self.canvasView.drawing.dataRepresentation()
-        self.viewModel?.updatePencilData(to: data, width: Int(self.canvasView.frame.width))
+        self.viewModel?.updatePencilData(to: data, width: Double(self.canvasView.frame.width))
     }
 }
